assignment to 'uint16_t *' from incompatible pointer type 'uint8_t *'

+++ user/source/handlers.c
@@ -12,7 +12,7 @@ extern uint8_t vADCConvertedArray_1[ui32RecBuffSize];
 extern xQueueSetHandle xDMAData;
 
 void DMA2_Stream3_IRQHandler(void) {
-	uint16_t* ui16RecBufferPoint;
+	uint8_t* ui16RecBufferPoint;
 	if (DMA_GetITStatus(DMA2_Stream3, DMA_IT_TCIF3)) {
 		DMA_ClearITPendingBit(DMA2_Stream3, DMA_IT_TCIF3);
 		if (DMA_GetCurrentMemoryTarget(DMA2_Stream3) == 0) {
